Rename gatsby-node-utils to gatsby-utils; remove unneeded TODO
